+++ integration_test/tests/index.ts
@@ -1,8 +1,10 @@
 import AdvertisingTest from './advertisingTest';
 import audioTrackTest from './audioTrackTest';
+import captionMetadataTest from './captionMetadataTest';
 import CaptionTest from './captionTest';
 import ErrorTest from './errorTest';
 import LoadingTest from './loadingTest';
+import MediaControlsTest from './mediaControlsTest';
 import metadataId3Test from './metadataId3Test';
 import PlaybackTest from './playbackTest';
 import UnloadingTest from './unloadingTest';
@@ -11,8 +13,10 @@ import videoQualityTest from './videoQualityTest';
 export default [
   AdvertisingTest,
   CaptionTest,
+  captionMetadataTest,
   ErrorTest,
   LoadingTest,
+  MediaControlsTest,
   metadataId3Test,
   PlaybackTest,
   UnloadingTest,